requirements.txt
feedparser==6.0.10
pytz==2023.3
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
feedgenerator==2.1.0
markdown==3.5.1
flask==2.0.1
gunicorn==20.1.0